silence size conversion warning under x64
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@17269 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/Utils/UnicodeUtils.cpp
@@ -150,7 +150,7 @@ CString CUnicodeUtils::UTF8ToUTF16 (const std::string& string)
         while (c != 0);
     }
 
-    return CString (result, target - result);
+    return CString (result, static_cast<int>(target - result));
 }
 
 CStringA CUnicodeUtils::ConvertWCHARStringToUTF8(const CString& string)
